Move SpectrumProcessor.py to filtering
