+++ matchms/Pipeline.py
@@ -9,7 +9,7 @@ from matchms.importing.load_spectra import load_list_of_spectrum_files
 from matchms.logging_functions import (add_logging_to_file,
                                        reset_matchms_logger,
                                        set_matchms_logger_level)
-from matchms.SpectrumProcessor import SpectrumProcessor
+from matchms.filtering.SpectrumProcessor import SpectrumProcessor
 from matchms.typing import SpectrumType
