+++ matchms/filtering/SpectrumProcessor.py
